requirements.txt
Django==5.0.2
python-dotenv==1.0.1
Pillow==11.2.1  # Latest stable version
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-allauth==0.61.1
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
google-auth==2.27.0 